+++ data/src/data_channel/mod.rs
@@ -118,9 +118,7 @@ impl DataChannel {
             })
             .marshal()?;
 
-            stream
-                .write_sctp(&msg, PayloadProtocolIdentifier::Dcep)
-                .await?;
+            stream.write_sctp(&msg, PayloadProtocolIdentifier::Dcep)?;
         }
         Ok(DataChannel::new(stream, config))
     }
@@ -286,13 +284,10 @@ impl DataChannel {
         };
 
         let n = if data_len == 0 {
-            let _ = self
-                .stream
-                .write_sctp(&Bytes::from_static(&[0]), ppi)
-                .await?;
+            let _ = self.stream.write_sctp(&Bytes::from_static(&[0]), ppi)?;
             0
         } else {
-            let n = self.stream.write_sctp(data, ppi).await?;
+            let n = self.stream.write_sctp(data, ppi)?;
             self.bytes_sent.fetch_add(n, Ordering::SeqCst);
             n
         };
@@ -305,8 +300,7 @@ impl DataChannel {
         let ack = Message::DataChannelAck(DataChannelAck {}).marshal()?;
         Ok(self
             .stream
-            .write_sctp(&ack, PayloadProtocolIdentifier::Dcep)
-            .await?)
+            .write_sctp(&ack, PayloadProtocolIdentifier::Dcep)?)
     }
 
     /// Close closes the DataChannel and the underlying SCTP stream.